+++ src/lib/api/transform.spec.ts
@@ -103,6 +103,7 @@ describe('transform', () => {
     assert.equal(result, expected);
   });
 
+
   it('should handle storage with url handle', () => {
     const storeAlias = 'https://test.com/file.js';
 
@@ -171,8 +172,30 @@ describe('transform', () => {
     };
 
     const result = transform(url, testConfig);
-    assert.equal(result, `${cdnUrl}/${url}`);
+    assert.equal(result, `${cdnUrl}/polaroid/${url}`);
   });
+
+  it('should handle store without params', () => {
+    const options = {};
+    const result = transform(url, { store: options});
+
+    assert.equal(result, `${cdnUrl}/store/${url}`);
+  });
+
+  it('should return correct store URL with "/" in path', () => {
+    const options = { filename: 'test.jpg' , path: 'test/path'};
+    const result = transform(url, { store: options});
+
+    assert.equal(result, `${cdnUrl}/store=filename:test.jpg,path:"test/path"/fakelink`)
+  });
+
+  it('should handle upper cased path', () => {
+    const options = { filename: 'test.jpg' , path: 'tEsT/path'};
+    const result = transform(url, { store: options});
+
+    assert.equal(result, `${cdnUrl}/store=filename:test.jpg,path:"tEsT/path"/fakelink`)
+  });
+
 
   describe('blackwhite', () => {
     it('should construct valid parameters', () => {